add post subscribe function
