+++ src/components/Subscribe/index.js
@@ -1,0 +1,1 @@
+export { default as Subscribe } from "./Subscribe";
