--- src/components/SubscribeBtn/index.js
@@ -1,1 +1,0 @@
-export { default as SubscribeButton } from "./SubscribeButton";
